+++ packages/framer-motion/src/motion/__tests__/transformTemplate.test.tsx
@@ -2,6 +2,7 @@ import { render } from "../../../jest.setup"
 import { motion } from "../../"
 import * as React from "react"
 import { frame } from "../../frameloop"
+import { nextMicrotask } from "../../gestures/__tests__/utils"
 
 describe("transformTemplate", () => {
     it("applies transformTemplate on initial render", () => {
@@ -18,7 +19,7 @@ describe("transformTemplate", () => {
         )
     })
 
-    it("applies updated transformTemplate", () => {
+    it("applies updated transformTemplate", async () => {
         const { container, rerender } = render(
             <motion.div
                 initial={{ x: 10 }}
@@ -40,6 +41,8 @@ describe("transformTemplate", () => {
                 }}
             />
         )
+
+        await nextMicrotask()
         expect(container.firstChild).toHaveStyle(
             "transform: translateY(20px) translateX(10px) translateZ(0)"
         )
@@ -74,6 +77,7 @@ describe("transformTemplate", () => {
             />
         )
         expect(container.firstChild).toHaveStyle("transform: translateY(20px)")
+
         rerender(<motion.div style={{ x: 20 }} />)
 
         await new Promise((resolve) => frame.postRender(resolve))
@@ -104,6 +108,7 @@ describe("transformTemplate", () => {
         const { container, rerender } = render(
             <motion.div transformTemplate={() => `translateY(20px)`} />
         )
+        await nextMicrotask()
         expect(container.firstChild).toHaveStyle("transform: translateY(20px)")
         rerender(<motion.div />)