+++ packages/framer-motion/src/projection/node/create-projection-node.ts
@@ -56,6 +56,7 @@ import { clamp } from "../../utils/clamp"
 import { steps } from "../../frameloop/frame"
 import { noop } from "../../utils/noop"
 import { time } from "../../frameloop/sync-time"
+import { microtask } from "../../frameloop/microtask"
 
 const transformAxes = ["", "X", "Y", "Z"]
 
@@ -555,6 +556,7 @@ export function createProjectionNode<I>({
         // Note: currently only running on root node
         startUpdate() {
             if (this.isUpdateBlocked()) return
+
             this.isUpdating = true
             this.nodes && this.nodes.forEach(resetRotation)
             this.animationId++
@@ -571,6 +573,7 @@ export function createProjectionNode<I>({
                 this.options.onExitComplete && this.options.onExitComplete()
                 return
             }
+
             !this.root.isUpdating && this.root.startUpdate()
             if (this.isLayoutDirty) return
 
@@ -658,7 +661,7 @@ export function createProjectionNode<I>({
         didUpdate() {
             if (!this.updateScheduled) {
                 this.updateScheduled = true
-                queueMicrotask(() => this.update())
+                microtask.read(() => this.update())
             }
         }
 
@@ -1056,9 +1059,7 @@ export function createProjectionNode<I>({
              * a relativeParent. This will allow a component to perform scale correction
              * even if no animation has started.
              */
-            // TODO If this is unsuccessful this currently happens every frame
             if (!this.targetDelta && !this.relativeTarget) {
-                // TODO: This is a semi-repetition of further down this function, make DRY
                 const relativeParent = this.getClosestProjectingParent()
                 if (
                     relativeParent &&